// src/app/models/signup.model.ts
export interface SignupModel {
  username: string;     
  password: string; 
  project: string;
  registrationNumber: string;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  dateOfBirth: string;
  motherName: string;
  nationality: string;
  city: string;
  cnic: string;
  cnicExpiry: string;
  passportNumber: string;
  passportExpiry: string;
  primaryWhatsApp: string;
  alternateMobile: string;
  mailingAddress: string;
  permanentAddress: string;
  agreeTerms: boolean;
}